Patch - added waiting screen for firebase petitions with animation spinner
