+++ MemoLab/Config/FirebaseAuth/UserAuthViewModel.swift
@@ -14,8 +14,10 @@ final class UserAuthViewModel: ObservableObject {
     @Published var hasError: Bool = false
     @Published var error: String = ""
     @Published var launchVerifyEmailModal: Bool = false
+    @Published var isWaitingResponse: Bool = false
     
     func signUp(with credentials: UserCredentials) async {
+        self.isWaitingResponse = true
         let result = await UserAuthRepository.signUp(with: credentials.email, and: credentials.password)
         switch result {
         case .success(let user):
@@ -25,34 +27,43 @@ final class UserAuthViewModel: ObservableObject {
             self.error = ""
             self.launchVerifyEmailModal = !user.emailIsVerified
             await self.sendVerificationEmail()
+            self.isWaitingResponse = false
         case .failure(let error):
             self.userAuth = nil
             self.userCanAccess = false
             self.hasError = true
             self.error = error.localizedDescription
             self.launchVerifyEmailModal = false
+            self.isWaitingResponse = false
         }
     }
     
     func sendVerificationEmail() async {
+        self.isWaitingResponse = true
         let result = await UserAuthRepository.sendVerificationEmail()
         switch result {
         case .success(_):
             self.launchVerifyEmailModal = true
+            self.isWaitingResponse = false
         case .failure(_):
             self.launchVerifyEmailModal = false
+            self.isWaitingResponse = false
         }
     }
     
     func isUserLoggedInAndVerified() async {
+        self.isWaitingResponse = true
         guard let user = UserAuthRepository.isUserLoggedIn() else {
+            self.isWaitingResponse = false
             return
         }
         self.userAuth = user
         self.userCanAccess = user.isAnonymous || user.emailIsVerified
+        self.isWaitingResponse = false
     }
     
     func signInAnonymously() async {
+        self.isWaitingResponse = true
         let result = await UserAuthRepository.signInAnonymously()
         switch result {
         case .success(let user):
@@ -61,16 +72,19 @@ final class UserAuthViewModel: ObservableObject {
             self.hasError = false
             self.error = ""
             self.launchVerifyEmailModal = false
+            self.isWaitingResponse = false
         case .failure(let error):
             self.userAuth = nil
             self.userCanAccess = false
             self.hasError = true
             self.error = error.localizedDescription
             self.launchVerifyEmailModal = false
+            self.isWaitingResponse = false
         }
     }
     
     func signIn(with credentials: UserCredentials) async {
+        self.isWaitingResponse = true
         let result = await UserAuthRepository.signIn(with: credentials.email, and: credentials.password)
         switch result {
         case .success(let user):
@@ -79,16 +93,19 @@ final class UserAuthViewModel: ObservableObject {
             self.hasError = false
             self.error = ""
             self.launchVerifyEmailModal = !user.emailIsVerified
+            self.isWaitingResponse = false
         case .failure(let error):
             self.userAuth = nil
             self.userCanAccess = false
             self.hasError = true
             self.error = error.localizedDescription
             self.launchVerifyEmailModal = false
+            self.isWaitingResponse = false
         }
     }
     
     func signOut() async {
+        self.isWaitingResponse = true
         let result = await UserAuthRepository.signOut()
         
         switch result {
@@ -96,6 +113,7 @@ final class UserAuthViewModel: ObservableObject {
             cleanAll()
         case .failure(let error):
             self.error = error.localizedDescription
+            self.isWaitingResponse = false
         }
     }
     
@@ -109,6 +127,7 @@ final class UserAuthViewModel: ObservableObject {
         self.userCanAccess = false
         self.cleanError()
         self.launchVerifyEmailModal = false
+        self.isWaitingResponse = false
     }
     
 }
